requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
openai>=1.0.0
sentence-transformers>=2.2.0
requests>=2.26.0
pillow>=9.0.0
markdown>=3.3.6
pandas>=1.3.0
duckdb>=0.9.0
whisper-ctranslate2==0.5.2  # Changed from >=2.0.0 to latest stable version
pytest>=7.0.0
pytest-asyncio>=0.16.0